curriculum-planner/requirements.txt
python-docx>=1.1.0
python-pptx>=0.6.23
lxml>=4.9.0
